New part to the projct
I added another file to keep the first’s week’s work intact. I have
tripped up on a couple of things over the weekend and haven’t gotten it
quite to work yet.  I got wicked sick Saturday night and part way
through Sunday and did not work on this as much as I wanted, I
apologize for that.  I will work on it today.